fix: split native spec by spec
